adding photos of Philip Eugene Buchanan
On 5/24/2025 after a very long search going back to 2013 when I began this family tree, I received photos of Philip Eugene Buchanan in the mail from someone I found who was in possession of some. On 5/27/2025, I scanned them in and added them here into the family tree. It is possible and maybe likely that these photos are the only remaining photos of Philip in existence.